requirements.txt
attrs==23.1.0
behave==1.2.6
certifi==2023.11.17
cffi==1.16.0
colorama==0.4.6
dnspython==2.4.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
parse==1.19.1
parse-type==0.6.2
pluggy==1.3.0
pycparser==2.21
pymongo==4.6.0
PySocks==1.7.1
pytest==7.4.3
python-dotenv==1.0.0
PyYAML==6.0.1
schedule==1.2.1
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
